+++ src/com/smartpark/models/Manager.java
@@ -1,7 +1,8 @@
 package com.smartpark.models;
+import java.math.BigDecimal;
 
 public class Manager extends Employee {
-    public Manager(int userId, String name, String contactNumber, String shift, double salary) {
+    public Manager(int userId, String name, String contactNumber, String shift, BigDecimal salary) {
         super(userId, name, contactNumber, "Manager", shift, salary);
     }
 }
