+++ src/com/smartpark/models/SecurityGuard.java
@@ -1,7 +1,8 @@
 package com.smartpark.models;
+import java.math.BigDecimal;
 
 public class SecurityGuard extends Employee {
-    public SecurityGuard(int userId, String name, String contactNumber, String shift, double salary) {
+    public SecurityGuard(int userId, String name, String contactNumber, String shift, BigDecimal salary) {
         super(userId, name, contactNumber, "Security Guard", shift, salary);
     }
 }
