+++ src/com/smartpark/services/SalaryService.java
@@ -3,6 +3,7 @@ package com.smartpark.services;
 import com.smartpark.dao.SalaryDAO;
 import com.smartpark.models.SalaryDisbursement;
 
+import java.math.BigDecimal;
 import java.sql.SQLException;
 import java.time.LocalDate;
 import java.util.List;
@@ -15,7 +16,8 @@ public class SalaryService {
         this.salaryDAO = new SalaryDAO();
     }
 
-    public void payEmployee(int employeeId, double amount, LocalDate paymentDate) throws SQLException {
+    // SIGNATURE UPDATED FOR BIGDECIMAL
+    public void payEmployee(int employeeId, BigDecimal amount, LocalDate paymentDate) throws SQLException {
         salaryDAO.recordSalaryPayment(employeeId, amount, paymentDate);
     }
